or#896: tag the live-shaped Stripe tests stripelive
TestLiveStripeCatalogAutoCreate* and TestLiveStripeExtrasListing carried no
build tag: they compiled into a default `go test ./...` and were held back only
by an env check, so they were neither evidence nor absent. They now use the
`stripelive` tag the live Stripe webhook tests already use.
