+++ pkg/service/catalog_extras_test.go
@@ -2,10 +2,8 @@ package service
 
 import (
 	"context"
-	"github.com/open-rails/openrails/internal/railresolve"
 	"net/http"
 	"net/http/httptest"
-	"os"
 	"strings"
 	"testing"
 	"time"
@@ -482,35 +480,5 @@ func TestComputeSolanaSunsetExtras(t *testing.T) {
 	}
 	if e.Label != "premium.usd.23000000.30" {
 		t.Errorf("label should be the price content key, got %q", e.Label)
-	}
-}
-
-// -- live smoke (read-only; opt-in) ------------------------------------------------
-
-// TestLiveStripeExtrasListing runs the extras-detection LISTING (GETs only;
-// nothing is ever archived) against a real Stripe account. Opt-in via
-// OPENRAILS_LIVE_STRIPE_KEY; use a TEST-mode key.
-func TestLiveStripeExtrasListing(t *testing.T) {
-	key := strings.TrimSpace(os.Getenv("OPENRAILS_LIVE_STRIPE_KEY"))
-	if key == "" {
-		t.Skip("set OPENRAILS_LIVE_STRIPE_KEY (a Stripe TEST key) to run the live read-only extras listing")
-	}
-	rails := railresolve.FixedSet{
-		"stripe": {Rail: models.RailStripe, Stripe: &config.StripeRailConfig{SecretKey: key}},
-	}
-	lister := &catalog.StripeCatalogService{Config: &config.Config{}, Rails: rails}
-	products, prices, err := fetchStripeCatalog(context.Background(), lister)
-	if err != nil {
-		t.Fatalf("live stripe listing: %v", err)
-	}
-	// Empty local snapshot: every remote object shows up classified ours/foreign.
-	extras := computeStripeExtras(products, prices, buildSnapshotFromRows(nil, nil))
-	t.Logf("live stripe account: %d products, %d prices, %d extras vs an empty catalog", len(products), len(prices), len(extras))
-	for _, e := range extras {
-		marker := "foreign"
-		if e.Owned {
-			marker = "openrails-marked"
-		}
-		t.Logf("  %s %s (%s) active=%v [%s]", e.ObjectType, e.ExternalID, e.Label, e.Active, marker)
 	}
 }
