+++ pkg/service/catalog_provider_stripe_live_test.go
@@ -1,3 +1,11 @@
+//go:build stripelive
+
+// Live Stripe catalog tests (or#896): they hit the REAL Stripe test account.
+// Tagged so they are evidence when a lane runs them and absent otherwise —
+// never silently compiled into `go test ./...` and skipped on an env check.
+//
+//	go test -tags=stripelive ./pkg/service/ -run TestLiveStripe -v
+
 package service
 
 import (
